requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
pydantic
redis
passlib[bcrypt]
PyJWT
cryptography
fakeredis
pytest
pytest-cov
pytest-asyncio
httpx
pydantic-settings
prometheus-client
email-validator
black
ruff
mypy
pre-commit
